Revert "AI model"
