+++ be/main.py
@@ -1,6 +1,6 @@
 import uvicorn
 from fastapi import FastAPI
-from src.routers import device, environment, adafruit, auth, ai
+from src.routers import device, environment, adafruit, auth
 from src.startup import seed_admin_user
 from fastapi.middleware.cors import CORSMiddleware
 
@@ -20,7 +20,7 @@ app.include_router(device.router)
 app.include_router(environment.router)
 app.include_router(adafruit.router)
 app.include_router(auth.router)
-app.include_router(ai.router)
+
 
 @app.get("/")
 def root():
