+++ be/requirements.txt
@@ -6,8 +6,4 @@ python-dotenv==1.0.1
 pydantic
 requests
 passlib[bcrypt]
-pandas
-scikit-learn
-joblib
-matplotlib
 bcrypt==3.2.0
