--- be/src/models/ai.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-class AIResponse(BaseModel):
-    pump: int
-    fan: int
-    
-class AIRequest(BaseModel):
-    temp: float
-    humid: float
