+++ be/src/routers/environment.py
@@ -7,9 +7,7 @@ from src.controllers.environment_controller import (
     get_hourly_environment_data,
     get_history_environment_data,
 )
-
-from src.models.environment import EnvironmentData, HistoryRequest, HourlyRequest, en_threshold
-
+from src.models.environment import EnvironmentData, en_threshold
 from datetime import datetime
 # Định nghĩa router với tiền tố "/environment"
 router = APIRouter(prefix="/environment", tags=["Environment"])
@@ -43,11 +41,8 @@ async def api_set_threshold(request: en_threshold):
 
 
 @router.get("/historyData")
-# async def read_history(start_day, end_day):
-async def read_history(r: HistoryRequest):
+async def read_history(start_day, end_day):
     """ API để lấy dữ liệu trung bình mỗi ngày theo region, trong khoảng tgian start_dat đến end_day"""
-    start_day = r.start_day
-    end_day = r.end_day
     if type(start_day) is str:
         start_day = datetime.fromisoformat(start_day)
     if type(end_day) is str:
@@ -61,10 +56,8 @@ async def read_history(r: HistoryRequest):
     return data
 
 @router.get("/hourlyData")
-# async def read_hourly(date=None):
-async def read_hourly(r: HourlyRequest):
+async def read_hourly(date=None):
     """date = None là lấy dữ liệu ngày hôm nay"""
-    date = r.date
     if type(date) is str:
         date = datetime.fromisoformat(date)
     data = get_hourly_environment_data(date)
